bug fix for ic in unit convertors

+++ src/unit_conversion/to_si.jl
@@ -1,14 +1,14 @@
 function pu_to_si!(data::Dict{String,Any},
     params::Dict{Symbol,Any}, nominal_values::Dict{Symbol,Any})
 
-    rescale_mass_flow = x -> x * nominal_values[:mass_flow]
-    rescale_mass_flux = x -> x * nominal_values[:mass_flux]
-    rescale_time = x -> x * nominal_values[:time]
-    rescale_pressure = x -> x * nominal_values[:pressure]
-    rescale_length = x -> x * nominal_values[:length]
-    rescale_density = x -> x * nominal_values[:density]
-    rescale_diameter = x -> x * nominal_values[:length]
-    rescale_area = x -> x * nominal_values[:area]
+    rescale_mass_flow = x -> Float64(x * nominal_values[:mass_flow])
+    rescale_mass_flux = x -> Float64(x * nominal_values[:mass_flux])
+    rescale_time = x -> Float64(x * nominal_values[:time])
+    rescale_pressure = x -> Float64(x * nominal_values[:pressure])
+    rescale_length = x -> Float64(x * nominal_values[:length])
+    rescale_density = x -> Float64(x * nominal_values[:density])
+    rescale_diameter = x -> Float64(x * nominal_values[:length])
+    rescale_area = x -> Float64(x * nominal_values[:area])
 
     function rescale_compressor_boundary_conditions!(type, value)
         @assert length(type) == length(value)
